Move Either to crate::utils
