+++ vm/src/lib.rs
@@ -77,6 +77,7 @@ pub mod slots;
 pub mod stdlib;
 pub mod sysmodule;
 pub mod types;
+pub mod utils;
 mod version;
 mod vm;
